chore: move all non adapter components out of UI
