+++ src/components/Common/FieldCaption/FieldCaption.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from 'react-i18next'
 import classNames from 'classnames'
-import { VisuallyHidden } from '../../VisuallyHidden'
+import { VisuallyHidden } from '../VisuallyHidden'
 import styles from './FieldCaption.module.scss'
 
 export interface FieldCaptionProps {
