+++ src/components/Common/UI/Checkbox/Checkbox.tsx
@@ -1,8 +1,8 @@
 import type { ChangeEvent } from 'react'
 import { useFieldIds } from '../hooks/useFieldIds'
-import { HorizontalFieldLayout } from '../HorizontalFieldLayout'
 import styles from './Checkbox.module.scss'
 import type { CheckboxProps } from './CheckboxTypes'
+import { HorizontalFieldLayout } from '@/components/Common/HorizontalFieldLayout'
 import IconChecked from '@/assets/icons/checkbox.svg?react'
 
 export const Checkbox = ({
