+++ src/components/Common/UI/CheckboxGroup/CheckboxGroupTypes.ts
@@ -1,5 +1,5 @@
 import type { FieldsetHTMLAttributes, Ref } from 'react'
-import type { SharedFieldLayoutProps } from '../FieldLayout/FieldLayoutTypes'
+import type { SharedFieldLayoutProps } from '@/components/Common/FieldLayout/FieldLayoutTypes'
 
 export type CheckboxGroupOption = {
   label: React.ReactNode
