+++ src/components/Common/UI/ComboBox/ComboBox.tsx
@@ -10,10 +10,10 @@ import {
 import { useTranslation } from 'react-i18next'
 import { useMemo } from 'react'
 import classNames from 'classnames'
-import { FieldLayout } from '../FieldLayout'
 import { useFieldIds } from '../hooks/useFieldIds'
 import styles from './ComboBox.module.scss'
 import type { ComboBoxProps } from './ComboBoxTypes'
+import { FieldLayout } from '@/components/Common/FieldLayout'
 import { useTheme } from '@/contexts/ThemeProvider'
 import AlertCircle from '@/assets/icons/alert-circle.svg?react'
 import CaretDown from '@/assets/icons/caret-down.svg?react'
