+++ src/components/Common/UI/DatePicker/DatePicker.tsx
@@ -15,10 +15,10 @@ import {
 import { useTranslation } from 'react-i18next'
 import classNames from 'classnames'
 import { parseDate } from '@internationalized/date'
-import { FieldLayout } from '../FieldLayout'
 import { useFieldIds } from '../hooks/useFieldIds'
 import styles from './DatePicker.module.scss'
 import type { DatePickerProps } from './DatePickerTypes'
+import { FieldLayout } from '@/components/Common/FieldLayout'
 import { useTheme } from '@/contexts/ThemeProvider'
 import CaretDown from '@/assets/icons/caret-down.svg?react'
 import CaretRight from '@/assets/icons/caret-right.svg?react'
