+++ src/components/Common/UI/NumberInput/NumberInput.tsx
@@ -1,11 +1,11 @@
 import { useMemo } from 'react'
 import { Group, NumberField as AriaNumberField } from 'react-aria-components'
 import classNames from 'classnames'
-import { FieldLayout } from '../FieldLayout'
 import { Input } from '../Input'
 import { useFieldIds } from '../hooks/useFieldIds'
 import styles from './NumberInput.module.scss'
 import type { NumberInputProps } from './NumberInputTypes'
+import { FieldLayout } from '@/components/Common/FieldLayout'
 import { useLocale } from '@/contexts/LocaleProvider'
 
 export function NumberInput({
