+++ src/components/Common/UI/Select/Select.tsx
@@ -11,9 +11,9 @@ import { useTranslation } from 'react-i18next'
 import { useMemo } from 'react'
 import classNames from 'classnames'
 import { useFieldIds } from '../hooks/useFieldIds'
-import { FieldLayout } from '../FieldLayout'
 import styles from './Select.module.scss'
 import type { SelectProps } from './SelectTypes'
+import { FieldLayout } from '@/components/Common/FieldLayout'
 import CaretDown from '@/assets/icons/caret-down.svg?react'
 import AlertCircle from '@/assets/icons/alert-circle.svg?react'
 import { useTheme } from '@/contexts/ThemeProvider'
